Reenable `readability-function-cognitive-complexity`
Default value of 25 is too low to pass since some functions are
substantially more complicated and we don't want to rewrite the
whole library to get it below a score of 25.

Disabling macros is necessary since the unit tests expand into
rather complicated functions but we don't want to be penalized for
how our test framework implements its macros.