Ensure all test data can be stringified
The CATCH_CONFIG_FALLBACK_STRINGIFIER macro expands into a function
that Catch2 will call when it cannot determine how to stringify a
given type. Because we define this to a nonsense value that will
never evaluate to a real function, this triggers a compiler error
when a given type cannot be stringified.

In doing so we had to fix many instances of testing types that lack
a string representation. In many cases this was as simple as including
a missing header but in some cases it required defining many extra
StringMaker specializations.

Now when any test fails, we're guaranteed to see a meaningful output
instead of "{?}".